% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/roxycode.R
\name{circadian_gene_plot}
\alias{circadian_gene_plot}
\title{circadian_gene_plot}
\usage{
circadian_gene_plot(gene_name, data1, tp1, group1, data2, tp2, group2)
}
\arguments{
\item{data1}{First data set.}

\item{tp1}{Time points of the first data set.}

\item{group1}{Group information of the first data set.}

\item{data2}{Second data set.}

\item{tp2}{Time points of the second data set.}

\item{group2}{Group information of the second data set.}

\item{gene_name:}{Name of the gene of interest, which exist in the row names of the data.}
}
\description{
This function performs plotting individual gene over time.
}
\author{
Viktorian Miok
}
