% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/roxycode.R
\name{circular_plot}
\alias{circular_plot}
\title{circular_plot}
\usage{
circular_plot(x, y, factors, color, leg_name, leg_col)
}
\arguments{
\item{x}{The log2 fold-changes of the dots to be plotted.}

\item{y}{The adjusted p-values of the dots to be plotted.}

\item{factors}{The variable which separate the dots on the peaces corresponding to the each time point.}

\item{color}{The color of the each dot in the circular plot.}

\item{leg_name}{Names of the different colors for the legend.}

\item{leg_col}{Color names for the legend.}
}
\description{
This function performs circular plot of the time points.
}
\author{
Viktorian Miok
}
