% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/roxycode.R
\name{data_summary}
\alias{data_summary}
\title{data_summary}
\usage{
data_summary(data, varname, groupnames)
}
\arguments{
\item{data}{Data set, numeric matrix of the values.}

\item{varname}{Name of the variable.}

\item{groupnames}{Name of the variable, for grouping.}
}
\description{
This function performs data summary per time point for each diet and tissue - function for plotting.
}
\author{
Viktorian Miok
}
