% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/roxycode.R
\name{dendro2sets}
\alias{dendro2sets}
\title{dendro2sets}
\usage{
dendro2sets(hc)
}
\arguments{
\item{hc}{Data set for conversion.}
}
\description{
format converting function - create set names from the dendrogram.
}
\author{
Viktorian Miok
}
