% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/roxycode.R
\name{dist.pear}
\alias{dist.pear}
\title{dist.pear}
\usage{
dist.pear(x)
}
\arguments{
\item{x}{Data set for distance calculation in heatmap}
}
\description{
This function performs distance calculation in heatmap.
}
\author{
Viktorian Miok
}
