% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/roxycode.R
\name{getFileSignF}
\alias{getFileSignF}
\title{getFileSignF}
\usage{
getFileSignF(filestyle)
}
\arguments{
\item{filestyle}{Data set for conversion.}
}
\description{
Extract the field separator character(FILE_SEP), the set of quoting characters(FILE_QUOTE), the character used for decimal points(FILE_DEC)
}
\author{
Viktorian Miok
}
