% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/roxycode.R
\name{hclust.ave}
\alias{hclust.ave}
\title{hclust.ave}
\usage{
hclust.ave(x)
}
\arguments{
\item{x}{Data set for distance calculation in heatmap.}

\item{method}{the agglomeration method to be used. This should be (an unambiguous abbreviation of) one of "ward.D", "ward.D2", "single", "complete", "average" (= UPGMA), "mcquitty" (= WPGMA), "median" (= WPGMC) or "centroid" (= UPGMC).}
}
\description{
This function performs hierahical clustering.
}
\author{
Viktorian Miok
}
