% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/roxycode.R
\name{pathway_association}
\alias{pathway_association}
\title{pathway_association}
\usage{
pathway_association(x_data, y_data, pathways, pathways_genes)
}
\arguments{
\item{x_data}{The response vector of the regression model.}

\item{y_data}{The part of the design matrix corresponding to the alternative hypothesis.}

\item{pathways}{The character vector of pathways names.}

\item{pathways_genes}{List object with all the gene names corresponding to the pathway names.}
}
\description{
This function performs association of selected pathways with peripheral data.
}
\author{
Viktorian Miok
}
