% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/roxycode.R
\name{plotUpset}
\alias{plotUpset}
\title{plotUpset}
\usage{
plotUpset(
  analyses,
  cut_off,
  col_interest,
  text.scale,
  matrix.color,
  main.bar.color,
  sets.bar.color,
  ...
)
}
\arguments{
\item{analyses}{Data set.}

\item{cut_off}{The number of intersections from each set (to cut off at) when aggregating by sets.}

\item{col_interest}{The columum number to be used.}

\item{text.scale}{Numeric, value to scale the text sizes, applies to all axis labels, tick labels, and numbers above bar plot.}

\item{matrix.color}{Color of the intersection points.}

\item{main.bar.color}{Color of the main bar plot.}

\item{sets.bar.color}{Color of set size bar plot.}
}
\description{
This function for plotting upset plot.
}
\author{
Viktorian Miok
}
