% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/roxycode.R
\name{rhythm_gene}
\alias{rhythm_gene}
\title{rhythm_gene}
\usage{
rhythm_gene(infile, filestyle, method, family, timepoints, knot, deg, ncpus)
}
\arguments{
\item{infile}{a character string. The name of input file containing time-series data.}

\item{filestyle}{a character vector. The data format of input file, must be "txt", or "csv".}

\item{method}{specify the method for detection of rhythmic signals, must be "tigaR", "JTK", "ARS", "LS".}

\item{family}{Character string. Either equal to "poisson", "zip" (zero-inflated Poisson), "nb" (negative binomial), or "zinb" (zero-inflated negative binomial): likelihood to be used..}

\item{knot}{number of knots to used for splines.}

\item{deg}{degree of splines to be used for splines.}

\item{ncpus}{Integer. The number of cpus to use for parallel computations.}
}
\description{
This function for detecting rhythmic gene expression over time.
}
\author{
Viktorian Miok
}
