% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/roxycode.R
\name{save_pheatmap_pdf}
\alias{save_pheatmap_pdf}
\title{save_pheatmap_pdf}
\usage{
save_pheatmap_pdf(obj, filename, width = 10, height = 10)
}
\arguments{
\item{obj}{pheatmap plot object.}

\item{filename}{location path and file name.}

\item{width}{width of the plot.}

\item{hight}{hight of the plot.}
}
\description{
This function for saving the pheatmap plot as pdf file.
}
\author{
Viktorian Miok
}
