% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/roxycode.R
\name{shrink_dge}
\alias{shrink_dge}
\title{shrink_dge}
\usage{
shrink_dge(
  dds,
  method = "shrinkbayes",
  contrast,
  family = "nb",
  num_gene = 100,
  maxiter = 3,
  ncpus = 2
)
}
\arguments{
\item{dds}{DESeqDataSet object. Result of DESeq function from DESeq2 package.}

\item{method}{Character. Indicated whether DESeq2 or ShrinkBayes package is employed for the analysis.}

\item{contrast}{Character string. Indicating the which groups to be compared.}

\item{family}{Character string. Either equal to "poisson", "zip" (zero-inflated Poisson), "nb" (negative binomial), or "zinb" (zero-inflated negative binomial): likelihood to be used.}

\item{num_gene}{Integer. The number of genes for ShrinkBayes method testing}

\item{maxiter}{Integer. Number of iteration of ShrinkBayes method.}

\item{ncpus}{Integer. The number of cpus to use for parallel computations.}
}
\description{
This function for differential expression analysis at each time point using schrinkage.
}
\author{
Viktorian Miok
}
